Add： 添加左右脑脑波节律实时数据

+++ EnterAffectiveCloudDemo/Report/ReportNewModel.swift
@@ -71,7 +71,7 @@ class ReportNewModel: NSObject {
     public var heartRateMin: Int?
     
     public var heartRateVariability: [Int]?
-    public var hrvAvg: Int?
+    public var hrvAvg: Float?
     
     public var attention: [Int]?
     public var attentionAvg: Int?
